Rename readers to gen_obs_writers
